Week 11/requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
python-jose[cryptography]
cryptography
python-multipart
pydantic[email]
